Project setup + prisma post learning
For learning reasons I didn't commit before, so it doesn't get messy, anyways, I'll be now posting the commits and prepare for my interviewer.

Things to do:
Test the seeding, migration is working.
Check and commit the actual system and interface with forms
Documentation
When it works, clone the repo and try to run it.

If you're reading this: I wish you have a great day!